# openapi.yaml
swagger: "2.0"
basePath: /
info:
  title: API demo with different backends
  version: 1.0.0
  description: API for managing users and cities
paths:
...
  /cities/{id}:
    get:
      summary: Get a city by ID
      operationId: "getCity"
      x-google-backend:
        address: https://new-go-app-pp46z4u6va-ew.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      parameters:
        - name: id
          type: integer
          in: path
          required: true
          description: ID of the city to retrieve
      responses:
        '200':
          description: Successful response
          schema:
            $ref: '#/definitions/City'
        404:
          description: User not found
          schema:
            $ref: '#/definitions/Error'
definitions:
...
    City:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
    Error:
      type: object
      properties:
        message:
          type: string